2FA works

+++ backend/data_controller/data/migrations/20250811115613_chat_data.js
@@ -4,13 +4,16 @@
  * @returns { Promise<void> }
  */
 export async function up(knex) {
-  await knex.schema.createTable('chat_messages', (table) => {
-    table.increments('id').primary();
-    table.integer('sender_id').unsigned().notNullable().references('id').inTable('credentialsTable').onDelete('CASCADE'); // If a user is deleted, their messages are also deleted.
-    table.integer('recipient_id').unsigned().notNullable().references('id').inTable('credentialsTable').onDelete('CASCADE');
-    table.text('message_text').notNullable();
-    table.timestamp('created_at').defaultTo(knex.fn.now());
-  });
+  const tableExists = await knex.schema.hasTable('chat_messages');
+  if (!tableExists) {
+    await knex.schema.createTable('chat_messages', (table) => {
+      table.increments('id').primary();
+      table.integer('sender_id').unsigned().notNullable().references('id').inTable('credentialsTable').onDelete('CASCADE'); // If a user is deleted, their messages are also deleted.
+      table.integer('recipient_id').unsigned().notNullable().references('id').inTable('credentialsTable').onDelete('CASCADE');
+      table.text('message_text').notNullable();
+      table.timestamp('created_at').defaultTo(knex.fn.now());
+    });
+  }
 };
 
 /**
